Some mild tweaking
The app now uses an ERB template to look a little more like a typical sinatra app.  It's had a mild facelift so you can tell what is the clever bit and what is just nasty HTML.

The detachme script has a working shebang and it's got a fork ribbon.

Rubymine complains less.